+++ app/src/test/java/woowacourse/movie/domain/policy/TimeDiscountPolicyTest.kt
@@ -1,9 +1,8 @@
-package woowacourse.movie.domain
+package woowacourse.movie.domain.policy
 
 import junit.framework.TestCase.assertEquals
 import junit.framework.TestCase.assertTrue
 import org.junit.Test
-import woowacourse.movie.domain.policy.TimeDiscountPolicy
 import java.time.LocalDateTime
 
 class TimeDiscountPolicyTest {
